Create theme folder. Change useTheme hook after font and new theme folder added
