--- src/helpers/constants/palette.ts
@@ -1,15 +1,0 @@
-export const darkTheme = {
-  primary: '#8CD6BD',
-  secondary: '#004761',
-
-  buttonBgColor: '#FFFFFF',
-  buttonTextColor: '#000000'
-}
-
-export const lightTheme = {
-  primary: '#004761',
-  secondary: '#8CD6BD',
-
-  buttonBgColor: '#000000',
-  buttonTextColor: '#FFFFFF'
-}
