+++ src/theme/fonts.ts
@@ -1,0 +1,3 @@
+export const fonts = {
+  inter_bold: 'Inter-Bold'
+}
