+++ src/theme/index.ts
@@ -1,0 +1,7 @@
+import { fonts } from './fonts'
+
+const theme = {
+  fonts
+}
+
+export default theme
