+++ src/theme/palette.ts
@@ -1,0 +1,20 @@
+const commonColors = {
+  white: '#FFFFFF',
+  black: '#000000'
+}
+
+export const darkTheme = {
+  primary: '#8CD6BD',
+  secondary: '#004761',
+
+  buttonBgColor: commonColors.white,
+  buttonTextColor: commonColors.black
+}
+
+export const lightTheme = {
+  primary: '#004761',
+  secondary: '#8CD6BD',
+
+  buttonBgColor: commonColors.black,
+  buttonTextColor: commonColors.white
+}
